requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
alembic
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
passlib[bcrypt]
python-jose
celery[redis]
redis
python-multipart
